restructure project
places source and header files within "include" & "source" folders
