+++ DX11-Base.vcxproj.filters
@@ -105,22 +105,19 @@
     <ClInclude Include="libs\MinHook\trampoline.h">
       <Filter>Resource Files\MinHook</Filter>
     </ClInclude>
-    <ClInclude Include="config.h">
+    <ClInclude Include="include\config.h">
       <Filter>Header Files\Class</Filter>
     </ClInclude>
-    <ClInclude Include="libs\utils\memory.h">
-      <Filter>Header Files</Filter>
-    </ClInclude>
-    <ClInclude Include="feature.h">
-      <Filter>Header Files</Filter>
-    </ClInclude>
-    <ClInclude Include="database.h">
+    <ClInclude Include="include\memory.h">
       <Filter>Header Files\Class</Filter>
     </ClInclude>
-    <ClInclude Include="ItemList.hpp">
+    <ClInclude Include="include\feature.h">
       <Filter>Header Files</Filter>
     </ClInclude>
-    <ClInclude Include="initialize.hpp">
+    <ClInclude Include="include\initialize.hpp">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="include\ItemList.hpp">
       <Filter>Header Files</Filter>
     </ClInclude>
   </ItemGroup>
@@ -185,13 +182,13 @@
     <ClCompile Include="libs\MinHook\trampoline.c">
       <Filter>Resource Files\MinHook</Filter>
     </ClCompile>
-    <ClCompile Include="config.cpp">
+    <ClCompile Include="src\memory.cpp">
       <Filter>Source Files\Class</Filter>
     </ClCompile>
-    <ClCompile Include="libs\utils\memory.cpp">
-      <Filter>Source Files</Filter>
+    <ClCompile Include="src\config.cpp">
+      <Filter>Source Files\Class</Filter>
     </ClCompile>
-    <ClCompile Include="feature.cpp">
+    <ClCompile Include="src\feature.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
   </ItemGroup>
