+++ include/feature.h
@@ -1,7 +1,6 @@
 #pragma once
 #include "include/Menu.hpp"
 #include "config.h"
-#include <Windows.h>
 
 void ESP();
